Duplicated code refactoring
