+++ src/app/component/search/search.component.spec.ts
@@ -32,8 +32,7 @@ describe('SearchComponent', () => {
             providers: [ 
                 { provide: AuthenticationService, useClass: AuthenticationServiceStub }, 
                 { provide: Router, useValue: routerSpy },
-                { provide: TripService, useClass: TripServiceStub },
-                { provide: GeoService, useClass: GeoServiceStub }
+                { provide: TripService, useClass: TripServiceStub }
             ],
             schemas: [ CUSTOM_ELEMENTS_SCHEMA ]
         })
