+++ src/app/testing/trip-service-stub.ts
@@ -75,4 +75,43 @@ export class TripServiceStub {
             }  
         };
     }
+
+    parseTrips(trips: any): Trip[] {
+        let jsTrip;
+        let users;
+        const parsed: Trip[] = [];
+        
+        for (const trip of trips) {
+            jsTrip = new Trip();
+
+            users = trip['users'];
+
+            jsTrip.users = users;
+            jsTrip.tripId = trip['tid'];
+            jsTrip.startLat = trip['start_lat'];
+            jsTrip.startLong = trip['start_lng'];
+            jsTrip.endLat = trip['end_lat'];
+            jsTrip.endLong = trip['end_lng'];
+            jsTrip.startTime = new Date(trip['start_time']);
+            jsTrip.endTime = new Date(trip['end_time']);
+            jsTrip.seats = trip['seats'];
+            jsTrip.smoking = trip['smoking'];
+            jsTrip.price = trip['price'];
+
+            // location strings
+            jsTrip.startLocation = `${trip.startLat || '-'},${trip.startLong || '-'}`;
+            jsTrip.endLocation = `${trip.endLat || '-'},${trip.endLong || '-'}`;
+
+            for (const uid in users) {
+                if (users[uid] === 'driving') {
+                    jsTrip.driver = uid.toString();
+                    break;
+                }
+            }
+
+            parsed.push(jsTrip);
+        }
+
+        return parsed;
+    }
 }
